Add MIT License
Add MIT License to the project. This file grants users permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software, provided that the copyright notice and license terms are included in all copies. This ensures compliance with open-source standards and provides clear legal permissions for usage and distribution.